Patch to allow merging of upstream reviews into devstack
The main goal is to facilitate easier testing of the Neutron DVR
code while it is still in development and not in the neutron
repository.

For that reason this patch should *NOT* be merged, but comments
are welcome since functionality like this might be good to have.

Things you should do before running stack.sh:

1) Copy your ssh key to ~stack/.ssh
2) Start the ssh authentication agent and add your key.  This will
   prevent you from being prompted for your password during cloning
   and merging:

	$ ssh-agent
	$ ssh-add

3) Add your gerrit id to local.conf:

	GERRIT_USER=john-doe

4) Add all upstream review id's to local.conf:

	NEUTRON_UPSTREAM_PATCHES="..."

   This has been pre-populated in samples/local.conf, but other
   review id's might be necessary as time goes on.

Possible ways to deal with merge failures:

1) Try running the merge script by hand, it's been provided separately:

	$ cd /opt/stack/neutron
	$ ~/devstack/merge-dvr-patchsets.sh $(review ids)

2) Provide comments on the upstream reviews so that they get
   merge conflicts fixed and/or rebased appropriately.

Future work on this patchset can enable configuration settings
in neutron config files to turn-on the DVR functionality.  That
work is TBD.

Change-Id: I669e054a8489fe7ac7f5c00df6535349d477e8c4