Add IPv6 support
A couple of things are covered in this patch
   -- the openstack management network can be either ipv4 or ipv6
   -- the tenant data network can be either ipv4 or ipv6

Define MGMT_NET with one of the two values 4 or 6 in your localrc
to indicate if you intend to run your management network as either
ipv4 or ipv6. Default value is 4

It's recommended to use host names to define HOST_IP, SERVICE_IP,
etc., in your localrc instead of IPv6 addresses. Those host names
should correspond to ipv6 addresses.

Define IP_VERSION with one of the three values 4, 6, or 4+6 in
your localrc to indicate if you intend to run your tenant data network
as either ipv4, ipv6, or dual stack respectively. Default value is 4.

A couple of variables should be defined in your localrc:
  - IPV6_PRIVATE_RANGE: the ipv6 prefix for your tenant network
  - IPV6_NETWORK_GATEWAY: the gateway IP with the same prefix

you may want to run a radvd service in your router namespace to advertise
default routes into your VMs. you can use this patch to achieve that:
https://review.openstack.org/#/c/102648

To use the IPv6 subnet API, you can define the foloowing in localrc:
  - IPV6_RA_MODE (with default as slaac)
  - IPV6_ADDRESS_MODE (with default as slaac)

Make sure that your VM image has dhcpv6 client enabled at bootup.
Or otherwise, you need to run it manually after the VM is booted.
It's recommended to run the lastest version of dnsmasq 2.68.
If you intend to enable internet access in your VM, make sure
your network node has ipv6 internet access, and the ipv6 prefix for
your tenant network is a GUA and routable.

Implements: blueprint ipv6-support
Change-Id: I848abf18e00e2a869697c5ef6366bc567dde448a