Fix installation of setuptools for fedora
On fedora 19 (not checked on other versions) with python-setuptools
package installed, pip is not correclty upgrading setuptools.
A false version of an egg-info is left in the systems, producing
errors when using setuptools. Running the update again -
"pip install -U setuptools" - fixes the issue. This can be
considered a workaround and the proper fix should go in pip.
An issue is opened there to fix it:
https://github.com/pypa/pip/issues/1917

Change-Id: I4a32d9f96a32910e2338071db522c869a96750ec