Update patch set 1
Patch Set 1:

What we did was get a Fedora minimal installation and use the same puppet recipes that openstack upstream uses to create the nodepool images, untill a few days ago, that was working fine. After that, when nodepool recreate the image based on the latest one (just like upstream do), we start to see this behaviour with setuptools.

Patch-set: 1