Update patch set 32
Patch Set 32:

@Dean, do you think this patch should be blocked on / rebased on top of your cinder multi-backend patch?

There may be someone willing to do the work you'd like to see here to get it in.  Can you elaborate a little more on what needs to be done.

I'm especially interested in getting this in as I'd like to help get some test configurations going in the gate using ceph as the storage backend.

Patch-set: 32