Update patch set 1
Patch Set 1:

(1 comment)

Wow...this makes me both nervous and excited.

One thing we need to be sure happens here before this can go in is that an out-of-the-box Neutron config 'just works' so whatever DevStack needs to do to make that happen should be included if it isn't already a default config.

Patch-set: 1
Label: Verified=0