Update patch set 1
Patch Set 1: -Workflow

Thanks for the review Dean and John!

Dean: This should just work, and if it doesn't, we'll fix this. The idea is that once we get the neutron full job running smoothly, perhaps we can consider this.

Removing WIP for now as well now that you've had a peek at this. :)

Patch-set: 1