Update patch set 35
Patch Set 35: Code-Review-1

Sébastien, Dean: On fedora, I seemed to be missing the 'redhat-lsb-core' package, which causes failure.  There doesn't seem to be another devstack dependency that pulls this in, so if it's not already installed then you can expect a error when executing lsb.  Can this be added to files/rpms/ceph ?

Patch-set: 35
Label: Code-Review=-1