Set postgresql max_connections=151
Now that we have multiple workers running by default
in various projects (nova/cinder/glance/trove), the
postgresql job is failing intermittently with connection
failures to the database.

The default max_connections for postgresql is 100 so here
we double that.

Note that the default max_connections for mysql used to
be 100 but is now 151, so this change brings the postgresql
configuration more in line with mysql.

Change-Id: I2fcae8184a82e303103795a7bf57c723e27190c9
Closes-Bug: #1338841
(cherry picked from commit 94c654ef37f6a0247a307578f3240f97201a3cba)