Update patch set 35
Patch Set 35: Code-Review+1

We are rapidly approaching a 1 year anniversary of the original blueprint that was superseded by Sébastien's.

https://blueprints.launchpad.net/devstack/+spec/ceph-storage-backend

I applaud Sébastien for having the patience keep with this for the last 7 months on his attempt to get this merged. A lot of us appreciate it as we have been using this unmerged patch to push openstack ceph support forward for almost a year.

Patch-set: 35
Reviewer: Gerrit User 7893 <7893@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1