TESTING: Add secondary mirror for ubuntu
We are seeing occasional rackspace apt mirror failures, to work around
this in case of a failure swap mirror with anl.gov.
Use anl.gov because it is in the US, has a 10GB network and is usually
up to date.

This workaround is needed because apt-get doesn't know how to gracefully
recover if the primary source vanishes.

This is for testing only a proper solution belongs in devstack-gate.
devstack-gate can run an apt-get update before entering devstack and run
this action if apt-get update has issues.

Testing what happens with a mirror that will return 404 errors.

Change-Id: I5e9fc1edab8f6531285b3ea02641aec9470cad4d