TESTING: Add secondary mirror for ubuntu
Use apt-add-repository to add a backup apt mirror.

We are seeing occasional rackspace apt mirror failures, to work around
this add as secondary apt-mirror. Use anl.gov because it is in the US,
has a 10GB network and is usually up to date.

This is for testing only a proper solution belongs in devstack-gate.

Testing what happens with a mirror that will return 404 errors.

Change-Id: I5e9fc1edab8f6531285b3ea02641aec9470cad4d