Update patch set 2
Patch Set 2: Code-Review-1

The code looks OK; but can you explain why this happens to verify this is the correct fix?  The original bug says "During install, the Trove guest image (ubuntu_mysql.qcow2) is uploaded to glance twice" but doesn't explain how, and I would not expect the init to run twice?  that might be a real bug

Do you mean during a regular install, or is this possibly when running under grenade?

Patch-set: 2
Label: Code-Review=-1