Update patch set 2
Patch Set 2:

To clarify the original problem: During a normal install, the trove guest image is uploaded to glance two times.

1. stack.rc adds TROVE_GUEST_IMAGE_URL to IMAGE_URLS. All images in the IMAGE_URLS list are uploaded to glance by stack.sh.
2. lib/trove init also uploads the TROVE_GUEST_IMAGE_URL to glance. 

We could choose not to add the url to IMAGE_URLS to begin with, but I prefer the approach of making lib/trove init more robust and have it add the image if it is missing.

Patch-set: 2