Update patch set 2
Patch Set 2:

True, I could drop the bit where we upload the image if it is not found. 

We still need the change to check for the image in glance and obtain the id, since it is used for the trove-manage calls. I don't see much reason /not/ to have a fall-back case if the image is not found, but if folks find that confusing we could remove it.

Patch-set: 2