Update patch set 2
Patch Set 2:

Thanks for fixing that.

My multi-node nova setup experience with devstack is small, I'm just trying to think this through.  You say that it's necessary to have this in mutli-node installs; thus if someone is doing the obvious thing of running a ./stack.sh with just n-cpu enabled; you're saying that it won't work unless they've also specified NOVA_VNC_ENABLED?

Because that sounds really confusing, especially as lots of people are starting to think about multi-node.  It feels like it would be best if we had something that worked more automatically in this case?

Am I mis-understanding?  Any ideas if we could auto-match some conditions to enable this?

Patch-set: 2
Label: Verified=0