Update patch set 2
Patch Set 2:

> thus if someone is doing the obvious thing of running a ./stack.sh with just n-cpu enabled; you're saying that it won't work unless they've also specified NOVA_VNC_ENABLED?

That is correct.  In order for horizon on a controller to be able to see VNC sessions and the like from separate controllers, those controllers need to have `vnc_enabled = True` in nova.conf as well as the proxying information present there (i.e., `vncserver_listen` and `vncserver_proxyclient_address` set).

The only way to do this currently is to enable n-novnc or n-xvnc on the compute node.  (In lib/nova, the only places vnc_enabled are set are both at ~line 505 and are currently dependent on the presence or lack of these services.)

To make this more concrete, local.confs for a compute node and a controller node can be found at http://paste.openstack.org/show/86945/ (these should work, but will not have VNC accessible through the dashboard).  Adding n-novnc to `$ENABLED_SERVICES` on the compute's local.conf will cause it to work, though it will now be running the extraneous service.  On the other hand, with my changes, if one adds `NOVA_VNC_ENABLE=True` to compute's local.conf, it will also work without enabling n-novnc.

Ideally, yes, we would check something more automatic for multi-node cases.  The only thing I can think of to check is `$MULTI_HOST`, and my worry there is that we would then set `vnc_enabled=True` and the like in too many cases.  Perhaps someone more familiar with the code could comment on this?

Patch-set: 2