Update patch set 2
Patch Set 2:

In that case, should we just turn this on by default for all compute nodes?  Or maybe, if required, invert the logic so you have to *disable* this if for some reason you don't want it?

Because it seems like in the most usual case, you always want it enabled, independent of running the proxies?

Patch-set: 2