Update patch set 1
Patch Set 1: Code-Review-1

Is there an upstream bugzilla to fix the problem? It seems weird that it would be set to a non working default. Once there is an upstream bug I'm ok with us landing the work around, but we should comment the work around in the template (so hopefully we can remove it later).

Patch-set: 1
Label: Code-Review=-1