Update patch set 2
Patch Set 2:

We are building CIRROS image for other platforms. In my case, for ppc64. For now, we need to rename the image name to cirros-xxxxx-x86_64, so we can use it correctly in devstack. This patch would avoid to rename it.

I don't mind which name should the variable have. Maybe another opinion on this would be appreciate.

Patch-set: 2