Support for qemu and libvirt custom git repositories
The user can now define the variables(local.conf):
- LIBVIRT_REPO and LIBVIRT_BRANCH for libvirt
- QEMU_REPO and QEMU_BRANCH for qemu

to enable the compilation and the installation of libvirt and
qemu from git repositories.

LIBVIRT_PY_REPO and LIBVIRT_PY_BRANCH are required when LIBVIRT_REPO
is set, because the installation of the libvirt-python's
dependencies from the OS package manager may conflict with
the compiled libvirt.

Change-Id: I89d20609fbfc8791892ab103c7b25c1334e3e94d
Signed-off-by: Michele Paolino <m.paolino@virtualopensystems.com>