Update patch set 2
Patch Set 2: Code-Review+2

The only way we are going to be able to set anything 'automatically' is to build a unified local.conf for multi-node that allows a compute node to peek into a controller node config to check for presence of either of the VNC proxies.

The MULTI_NODE var is poorly named and is not useful as a general indicator, it affects how nova-network is configured and may not necessarily be set in all setups.

The unified config is a long way off, if it is ever implemented, so the next least-bad solution is to add yet another config knob.  This default, to not change current behaviour, is correct.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2