Update patch set 41
Patch Set 41: Code-Review-1

I'd like to see the tail patch on this which actually tries to test it on trusty pass before we land this. Because if only the fedora side is working, I don't think it should be in devstack.

Patch-set: 41
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1