An initial implementation of a 1.0 based messaging driver
This change adds a new RPC backend as an option for devstack.  This
new RPC backend is the Qpid Dispatch Router.  It is a lightweight
message router that supports the AMQP 1.0 protocol.  For more
information, see
http://qpid.apache.org/components/dispatch-router/index.html

This new backend is necessary in order to test the new AMQP 1.0
oslo.messaging transport as described in the blueprint.

Change-Id: Ib8dea59922844e87d6c947b5dca557f5b5fc1160
Implements: blueprint amqp10-driver-implementation