Allow tests of the AMQP 1.0 based olso.messaging driver
This change adds a configuration option that sets the messaging
protocol that is used by the RPC backend and client.

Some brokers can support different kinds of 'on the wire' messaging
protocols.  Qpid, for example, supports both AMQP 0-10 (the default),
and AMQP 1.0.  Use the RPC_MESSAGING_PROTOCOL configuration variable
to override the default protocol for those brokers that support
multiple protocol options.

This new option is necessary in order to test the new AMQP 1.0
oslo.messaging transport as described in the blueprint.

Change-Id: Ib8dea59922844e87d6c947b5dca557f5b5fc1160
Implements: blueprint amqp10-driver-implementation