Update patch set 1
Patch Set 1: Workflow-1

OK, so I managed to get this back on Trove's radar and it was discussed in this past weeks meeting.  Seems to be a planned fix at the API level that should work around this.  Marking this as a WIP until that lands.  https://bugs.launchpad.net/trove/+bug/1333852

Patch-set: 1
Label: Workflow=-1