Update patch set 1
Patch Set 1: Code-Review-1

I'd actually started on this but you can have it.

The problem here is that you don't want to use the type (unless you change some other things in the code).  The new model goes through and creates a type for each backend, so for example if you do:

CINDER_ENABLED_BACKENDS=lvm:lvmdriver-1,lvm:lvmdriver-2

It will get the first one, but fail because it will try and create another type for lvmdriver2 with the same name (lvm).

Patch-set: 1
Label: Code-Review=-1