fix service stopping when screen is not used
In order to stop services in the non screen case we need to kill by
process group instead of parent process which is often gone by the
time we get around to trying to kill things in grenade.

Adjust unstack.sh to do this in the non screen case for testability.

Change-Id: If3952320d3f4b9e835c1ed5f9b538557fc85e692