Update patch set 1
Patch Set 1:

It seems to me, especially in light of enikanorov's comments, that the approach taken here doesn't work.  I would suggest potentially moving the config into start_neutron_agents(), or to be even cleaner, putting a separate config step before the start that contains only the important changes.  This makes the change more fine-grained, which should be easier to reason about as well.

Patch-set: 1