Update patch set 1
Patch Set 1: Code-Review-1

This change feels a bit heavy-handed to me, but that might be because I don't fully understand the issue.  In line with what Dean said, it would be nice to use some test to verify that everything is good to go before trying it.

I also realize that I may be taking up a not particularly worthwhile cause, but sleep is not a synchronization primitive.

Patch-set: 1
Reviewer: Gerrit User 7175 <7175@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1