Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I'm okay with this in principle, but I don't think it's broad enough.  In particular, there are other images we download (e.g., Fedora).  I think a better way to do this kind of change might be to add an IMAGE_ARCH (or some such) configuration, and then download all images based on that.

Patch-set: 4