Update patch set 2
Patch Set 2:

So I am missing something in trying to re-create the scenario where the pid is wrong. I can do it if I intentionally background the command:

 screen_it screen_it fake-service "cd $TOP_DIR; $TOP_DIR/tests/fake-service.sh &"

But we shouldn't be doing that and I don't see it in the current DevStack tree.  So it must be that some service(s) is daemonizing itself?  All good daemons have a -d or similar to stay in the foreground, that's what we should be doing here. _OR_ teach the daemon or start_XXXX service function to properly get the pid file in place, because all good daemons also drop a pid somewhere, right?

Now in the case of the above don't-do-that backgrounded command, it looks like the pgid for all of the children is still the now gone parent, aka process group leader.   maybe pkill -g $PID is enough?  It doesn't care if the process itself is there, it just kills everything with that as the pgid.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0