Update patch set 42
Patch Set 42: Code-Review+1

I just want to clarify ... are you guys saying that you want to see devstack running successfully, or also that you want to see all of tempest passing first?

I totally agree that we should see devstack working successfully.  However, I would argue against waiting for the test fixes.

This is a large patch and has been maintained for a long time.  It would be nice to not have to keep rebasing this.  Ceph is obviously already used quite heavily with OpenStack.  It would be really nice to be able to just land this and be able to iterate on fixing the known test failures.

Landing this will actually help with iterating on the test failures.  My plan was to enable an experimental job that could be run for patches intended to fix these issues.

Patch-set: 42
Reviewer: Gerrit User 1561 <1561@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1