Update patch set 42
Patch Set 42: Code-Review+2

I have had success running this on Trusty, but not with Tempest, just running basic stuff by hand.  Since merging this now doesn't appear to break configurations that leaves Ceph disabled, I'm OK with pushing it on in.  If as Russell mentions, having this merged is helpful to resolving the other issues, then lets do it.

And as the one who drug his feet for much of the life of this review I'm not going to insist on a fast resolution to passing Tempest, but I do think there are others who would feel that way.

Patch-set: 42
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2