Update patch set 4
Patch Set 4:

(2 comments)

I'm forcing internal communication through the proxy to exercise that code. In a real deployment this is likely how it would work as at least some of these services would be on different machines.

I can try to split it out into an SSL and TLS-proxy patch but there is so much overlap it would be rather difficult to do in any sort of clean way and has the potential to make rebasing even more difficult.

Patch-set: 4