Update patch set 3
Patch Set 3:

Sean, this patch is specific to Nova. Patching Devstack to support configuring the number of workers for Cinder and Glance can be done, but that would be another patch. As far as I know, Devstack is only hardcoding the number of works (4) for Nova, and that was my primary goal with this patch to fix it.

Patch-set: 3
Reviewer: Gerrit User 6469 <6469@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0