Update patch set 2
Patch Set 2:

Alon,  the CINDER_DRIVER configuration mechanism and the cinder_plugins directory will be gone soon after DevStack's stable/juno branch is cut.  The change to the multiple backend model had no bearing on the drivers themselves, just on how the Cinder configuration is built.  If your driver can't work in that environment it would be up to you to either detect that situation and print a warning or abort the stack.sh run.

Patch-set: 2