Work around pip installed pkgs on Rackspace image
The upstream rackspace image has a bunch of pip installed packages as
cloud-init was installed via pip due to a lack of available system
packages.  This can break further system package installs, such as
markdown, which fails with

---
 Error unpacking rpm package python-markdown-2.4.1-1.el7.noarch
  error: unpacking of archive failed on file
   /usr/lib/python2.7/site-packages/Markdown-2.4.1-py2.7.egg-info: cpio: rename
---

Because that is a directory for the pip-installed package, and a file
in the RPM

Remove all pip installed packages on rackspace images before we start
to work around this.  I have filed an upstream issue with Rackspace
(ticket-id 140804-ord-0000134) and the issue is being worked on.

Change-Id: Id12d175143ed3b8e024d057d65fa67505c08042a