Update patch set 2
Patch Set 2: Code-Review-1

Since this is a hack-around for brokenness somewhere else I'd prefer it go into fixup_stuff.sh.  That way it is a) isolated and b) available outside stack.sh, which is useful for preparing a clean env for a snapshot.

Also, is there a way to detect the Rax image?  Or detect the brokenness so it doesn't trigger on every stack.sh run?  I wouldn't be thrilled to do this every time if I happened to do dev on rhel7.

Patch-set: 2
Label: Code-Review=-1