Update patch set 4
Patch Set 4:

The change in the latest patchset is to allow marconi to run on devstack in distros that don't yet have Mongo >2.2. With the change suggested by Sean, we'll limit marconi devstack to Trusty (and possibly Fedora 20?). This is fine as far as gating is concerned, but will impact users who want to run marconi+devstack on other distros.
@dtroyer & @sdague, Thoughts?

Patch-set: 4