Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I think you've got a problem with the swift post command.

Also, I don't know how many people are using DevStack to build stand-alone Swift installs without Keystone (we've considered that config in the past), but this might not work in that configuration.  It's not a blocker for me since I don't think it is used much.

Patch-set: 1
Label: Code-Review=-1