Update patch set 2
Patch Set 2:

lib/ironic relies on lib/neutron to get all networking setup ahead of time.  Note that there is already a special case for is_baremetal there, which will ideally be removed in the near future.  This is a much smaller special casing IMHO.  An alternative would be to modify the tenant network from lib/ironic and flip the shared bit, but I don't know enough about Neutron to know if that is something that can be done after-the-fact.

Patch-set: 2
Label: Verified=0