Update patch set 2
Patch Set 2:

IIRC the Python projects generate AUTHORS as part of the release process.  We don't do releases.  I agree WRT email addresses, but I know mine is already all over the place.

If we think we need an AUTHORS file it could very easily be added to the rest of the static HTML generated for devstack.org in tools/build_docs.sh.

I don't like the idea of having a generated file in the repo, and that has been hashed out elsewhere already and I'm happy to follow that precedent.  Personally I'm just as happy to let it go away.

Patch-set: 2