Update patch set 2
Patch Set 2:

That is not correct, it does nothing.  There are files in files/apts and files/rpms that match up to service names.  If you have a prereq that only applies to q-l3 then it should be listed in files/apts/q-l3 and files/rpms/q-l3.

If the files do not exist, create them.

Patch-set: 2