Update patch set 2
Patch Set 2: Code-Review+1

The devstack gate is meant to test devstack as developers use it.  Part of the frustration that we've seen in the past few cycles is due to the gate configuration drifting from the standard devstack configuration.  Therefore, we are taking a harder line on ensuring that devstack works as intended from scratch in as close to the default configuration as possible.  If devstack needs to modify the firewall configuration for openstack to work correctly, it should.  But we should not have a gate configuration where all of the "correct" options are set which differs from the default configuration which is known not to work (or work well).

I believe that using iptables instead of firewalld was an option; perhaps devstack could handle making that transition?

Regardless, Devstack does all kinds of horrible things to a machine in order for openstack to run.  I don't think any of them would be acceptable on a developer workstation; rather it should be run on a dedicated vm, or in an environment appropriately externally firewalled.

Patch-set: 2
Label: Code-Review=+1