Update patch set 2
Patch Set 2: Code-Review-1

I think I'd rather see something simpler, just a FIREWALLD_DISABLE[0] or something that defaults to unset, and when True simply stops the firewalld service or removes it if there are things that might restart it later.

devstack-gate can set it True to get the benefit.

[0] I don't really care about the name but it should default to being unset and make sense when set...in my mind unset == False.

BTW, the changes to the prereq files don't change anything except cache the firewalld package on distros that wouldn't have it by default otherwise...why install it just to turn it off?

Patch-set: 2
Label: Code-Review=-1