Update patch set 2
Patch Set 2:

@Daniel ; I hear your complaint that it shouldn't disable your specific firewall rules, but in reality devstack is taking over the host completely

additionally, firewalld is not part of the default f20 cloud image; it appears to be something extra added by the rax f20 image

there seems to be no benefit to workarounds a) outside devstack and b) that will be rarely used.

BTW, the change does appear to drop the run time from about 1hour to about 30 minutes; in-line with Ubuntu.

Patch-set: 2
Label: Verified=0