Halt if firewalld found on Fedora 20
firewalld interacts badly with the libvirt on f20, causing slow-downs
so great that it can timeout the gate.

To avoid people having to debug this problem over and over, we stop
the devstack install.  We do not simply auto-remove firewalld, as some
people might like to ensure they have another firewall solution for
their devstack machine.

Change-Id: I5252a12223a35f7fb7a4ac3c58aa4a3cd1bc4799