Update patch set 4
Patch Set 4: Code-Review-1

Seriously no, just no. Disabling the firewall on a users machine is just not an acceptable thing todo, especially for a bug that is really not an issue for developers. Only the gate systems are majorly impacted by this, and the performance problem doesn't exist in newer versions of libvirt either.

Patch-set: 4
Label: Code-Review=-1