Update patch set 4
Patch Set 4:

> - this is guarded as f20 only; when it comes time to switch to f21 it will use the status quo

As someone running F20 it is of no consolation once you've exposed my machine countless malicious attackers on the network

> - if firewalld is not removed/disabled by default, every developer is left to debug
>  why tempest/vms are running so slowly -- how else do they know to turn on the magic
>  flag?.  unless we do it by default (this change).  I *hate* spending hours debugging
>  a problem, only to find it's a known issue and I just needed to turn this or that
>  hidden knob 

Out of the total number of people running devstack, I venture to suggest that the percentage
who also run tempest is in single figures. The severity of the problem of slow VMs under
tempest does not justify making *every* single developers' machines more insecure by disabling
their firewall.

> - just before this we turn off selinux (one day I hope to get it on, lower priority
> than getting the job working) and then we proceed to destroy the system with pip. 
> devstack never agreed to keep your firewall for you, and given it destroys everything
> else, this is not out of character.

Disabling SELinux is clearly not great but is a much less severe action than disabling
the firewall IMHO. Yes, devstack does a lot of changes to the make but I would not
characterize that as "destroying everything". After running devstack I expect that
the machine is broadly as functional as before, with the exception that more software
is running. Some things like apache / mysql have been screwed around with, but that is
far from destroying the machine.  I certainly do not expect that devstack opens up the
machine to network attacks on any listening services.

> - out of 3 images (upstream, hp, rax), only rax has firewalld turned on.  it's not
> the common case.  the others boot with nothing

What is in the cloud images doesn't concern me here as that is not what most
developers using devstack are running. Developers run devstack on laptops /
servers they own with a "standard" Fedora installation which absolutely
has firewalld as the default integral part of the OS.

> - as an end-user there is an argument that this should be fixed upstream. 
> you're upstream, so I defer to you on that :)

As I mentioned already this is already fixed upstream and in Fedora 21 and
in the virt-preview add-on repo for Fedora 20. It is too risky to include
the change in standard Fedora 20 repos.

Patch-set: 4