Update patch set 4
Patch Set 4:

The virt-preview repository is a rebuild of the packages from current rawhide with no QA testing applied. Stuff goes in there with little-to-no user testing. It is intended as a way for people who want to be able to test functionality from future Fedora without upgrading their entire distro to rawhide. It is not something you want to unconditionally enable because most OpenStack developers don't want to be used as guinea pigs for as yet untested libvirt/qemu packages - they just want to focus on getting their own OpenStack work right.

Patch-set: 4