Update patch set 4
Patch Set 4: Code-Review+2

Daniel, running DevStack on a machine you care about is not a supported configuration. As you have noted publicly a number of times it makes assumptions about what it can do to its host and does them.  Joshua Harlow has a radically different approach which should be friendlier to 'pet' systems.

> Out of the total number of people running devstack, I venture to suggest that the percentage who also run tempest is in single figures. The severity of the problem of slow VMs under tempest does not justify making *every* single developers' machines more insecure by disabling their firewall.

I'll suggest that the percentage of developers running DevStack on Fedora is smaller than you imagine. This is only a problem for devs running F20 on a machine that needs to protect itself from a hostile network without help.  firewalld may be that help, but then you are running an app on your firewall.  An app which itself takes great liberties with iptables directly.

Sadly, security and devstack do not belong in the same sentence.  I honestly hate that.  We do occasionally take steps to improve that situation.  But I assure you it will be a long time (possibly never) before I'll ever suggest that a DevStack build is the right model to secure OpenStack.  DevStack is a 20lb sledge hammer and security is finish nails.

> Developers run devstack on laptops / servers they own with a "standard" Fedora installation which absolutely has firewalld as the default integral part of the OS.

That sounds like what happens at your job, but I haven't seen that situation yet first hand.  Developers run DevStack in virtual machines, either locally on laptops/servers they own or in some cloud somewhere (including AWS).

Patch-set: 4
Label: Code-Review=+2