Update patch set 4
Patch Set 4:

There are certainly things devstack does to a system which are not nice, but it is the only supported option we have for openstack development whether or not it is intended for "pet" machines at all. This is not an argument for making devstack ever more hostile to people's machines. Also note that running with firewalld is an supported scenario in production deployments on Fedora & RHEL, so unconditionally disabling this means devs won't be exercising the codepath that real deployments will. So I really fail to see any credible justification for why this needs to be inflicted on every developer instead of simply temporarily disable in the gate CI config as proposed earlier.

Patch-set: 4