Update patch set 5
Patch Set 5: Code-Review-1

firewalld is a valid solution to use in production deployments of OpenStack on Fedora & RHEL and as such developers need to be able to use this scenario during development. Blocking use of devstack when firewalld is running prevents that, even if they don't care about the performance hit because they're not using tempest at all or have a newer libvirt installed.

How about at the end of running devstack, after we print out the address of horizon, we do this:

  if ...firewalld is running... and libvirt is less than 1.2.4...

     echo "WARNING: using firewalld in combination with libvirt < 1.2.4 has a major performance penalty on VM startup"
     echo "WARNING: recommend updating to newer libvirt or disabling firewalld as a workaround"
  fi

that will make it pretty obvious to developers without blocking valid usage scenarios for them

Patch-set: 5
Label: Code-Review=-1